+++ sdk/storage/azure-storage-blob/tests/conftest.py
@@ -29,6 +29,11 @@ def add_sanitizers(test_proxy):
     add_general_regex_sanitizer(regex=tenant_id, value="00000000-0000-0000-0000-000000000000")
     add_header_regex_sanitizer(key="Set-Cookie", value="[set-cookie;]")
     add_header_regex_sanitizer(key="Cookie", value="cookie;")
+    add_header_regex_sanitizer(key="x-ms-session-token", value="Sanitized")
+    add_general_regex_sanitizer(
+        regex=r"<SessionToken>[^<]*</SessionToken>", value="<SessionToken>Sanitized</SessionToken>"
+    )
+    add_general_regex_sanitizer(regex=r"<SessionKey>[^<]*</SessionKey>", value="<SessionKey>U2FuaXRpemVk</SessionKey>")
     add_oauth_response_sanitizer()
 
     add_header_regex_sanitizer(key="x-ms-copy-source-authorization", value="Sanitized")